RUE-854: split control-flow semantic analysis
